+++ src/cli/commands/install.rs
@@ -7,13 +7,14 @@ use crate::{
     cli::{
         commands::HandleCommand,
         display::{logging::error, not_found},
+        parameter_checks,
     },
     config::Config,
     installer::{InstallType, Installer, InstallerOptions, types::OptionalPackageId},
     platforms::Target,
     register::package_register::PackageRegister,
     repositories::manager::RepositoryManager,
-    utils::{parameter_checks, unwrap_or_exit::UnwrapOrExit},
+    utils::unwrap_or_exit::UnwrapOrExit,
 };
 
 /// Installs the specified packages, if a version is given that version will be installed,